+++ api/src/main/scala/dx/api/DxApi.scala
@@ -814,6 +814,8 @@ case class DxApi(version: String = "1.0.0", dxEnv: DXEnvironment = DXEnvironment
   def downloadFile(path: Path, dxfile: DxFile, overwrite: Boolean = false): Unit = {
     def downloadOneFile(path: Path, dxfile: DxFile): Boolean = {
       val fid = dxfile.id
+      val fileObj = path.toFile
+      val alreadyExists = fileObj.exists()
 
       try {
         // Use dx download. Quote the path, because it may contains spaces.
@@ -833,6 +835,10 @@ case class DxApi(version: String = "1.0.0", dxEnv: DXEnvironment = DXEnvironment
       } catch {
         case e: Throwable =>
           logger.traceLimited(s"error downloading file ${dxfile}", exception = Some(e))
+          // the file may have been partially downloaded - delete it before we retry
+          if ((overwrite || !alreadyExists) && fileObj.exists()) {
+            fileObj.delete()
+          }
           false
       }
     }
@@ -841,6 +847,7 @@ case class DxApi(version: String = "1.0.0", dxEnv: DXEnvironment = DXEnvironment
     if (dir != null && !Files.exists(dir)) {
       Files.createDirectories(dir)
     }
+    // we rely on the fact that exists() exits as soon as it encounters `true`
     val success = Iterator
       .range(0, DownloadRetryLimit)
       .exists { counter =>
@@ -880,7 +887,7 @@ case class DxApi(version: String = "1.0.0", dxEnv: DXEnvironment = DXEnvironment
 
   // Upload a local file to the platform, and return a json link.
   // Use 'dx upload' as a separate process.
-  def uploadFile(path: Path, destination: Option[String] = None): DxFile = {
+  def uploadFile(path: Path, destination: Option[String] = None, wait: Boolean = false): DxFile = {
     if (!Files.exists(path)) {
       throw new AppInternalException(s"Output file ${path.toString} is missing")
     }
@@ -888,12 +895,10 @@ case class DxApi(version: String = "1.0.0", dxEnv: DXEnvironment = DXEnvironment
     def uploadOneFile(path: Path): Option[String] = {
       try {
         // shell out to dx upload. We need to quote the path, because it may contain spaces
-        val dxUploadCmd = if (destination.isDefined) {
-          s"""dx upload "${path.toString}" --destination "${destination.get}" -p --brief"""
-        } else {
-          s"""dx upload "${path.toString}" --brief"""
-        }
-        logger.traceLimited(s"--  ${dxUploadCmd}")
+        val destOpt = destination.map(d => s"""--destination "${d}" -p""").getOrElse("")
+        val waitOpt = if (wait) "--wait" else ""
+        val dxUploadCmd = s"""dx upload "${path.toString}" --brief ${destOpt} ${waitOpt}"""
+        logger.traceLimited(s"CMD: ${dxUploadCmd}")
         SysUtils.execCommand(dxUploadCmd) match {
           case (_, stdout, _) if stdout.trim.startsWith("file-") =>
             Some(stdout.trim())
@@ -925,11 +930,11 @@ case class DxApi(version: String = "1.0.0", dxEnv: DXEnvironment = DXEnvironment
       .getOrElse(throw new Exception(s"Failure to upload file ${path}"))
   }
 
-  def uploadString(content: String, destination: String): DxFile = {
+  def uploadString(content: String, destination: String, wait: Boolean = false): DxFile = {
     // create a temporary file, and write the contents into it.
     val tempFile: Path = Files.createTempFile("upload", ".tmp")
     silentFileDelete(tempFile)
     val path = FileUtils.writeFileContent(tempFile, content)
-    uploadFile(path, Some(destination))
+    uploadFile(path, Some(destination), wait = wait)
   }
 }