+++ api/RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # dxApi
 
+## 0.2.0 (2021-04-23)
+
+* Adds option to wait for upload to `DxApi.uploadFile`
+* Allow caching file folder independenly of `describe()`
+
 ## 0.1.16 (2021-04-19)
 
 * Adds 'state' field to `DxFileDescribe`